moved redshift tests into one file and added test for partitions
